Fix logic for profile page image, UI for delete and cancel
